// src/types.ts
export type DebugOptions = {
	gridColumns?: number;
	accent?: string;
	panel?: boolean;
	shortcuts?: boolean;
	prefix?: string;
};

export type OverlaySet = {
	content: HTMLElement;
	paddingTop: HTMLElement;
	paddingRight: HTMLElement;
	paddingBottom: HTMLElement;
	paddingLeft: HTMLElement;
	marginTop: HTMLElement;
	marginRight: HTMLElement;
	marginBottom: HTMLElement;
	marginLeft: HTMLElement;
};